frontend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
cryptography>=41.0.0
